Refactoring group api into user
