+++ users/routes/user-routes.js
@@ -1,7 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const bcrypt = require('bcrypt');
-const { User, Statistics } = require('../services/user-model');
+const { User, Statistics, Group, UserGroup } = require('../services/user-model');
+const {  } = require('../services/user-model');
+
 
 //User internal routes
 const apiRoutes = require('../routes/user-routes-api');
@@ -89,6 +91,94 @@ router.post('/logout', async (req, res) => {
 router.get('/session', async (req, res) => {
     res.json({ session: req.session });
 });
+
+
+// Getting the list of groups in the database
+router.get('/list', async (req, res) => {
+    try {
+
+        const username = req.query.username;
+
+        // If the user is null or undefined (no one is logged, return all groups)
+        if (username === null || username === undefined) {
+          const allGroups = await Group.findAll({ order: [['name', 'ASC']] });
+          const groupsJSON = await Promise.all(allGroups.map(async (group) => {
+              const userCount = await UserGroup.count({
+                  where: {
+                      groupName: group.name
+                  }
+              });
+              return {
+                  name: group.name,
+                  isMember: false,
+                  isFull: userCount === 20
+              };
+          }));
+          return res.json({ groups: groupsJSON });
+      }
+
+        // If someone is logged, return the groups indicating which one the user has joined
+        const userGroups = await UserGroup.findAll({
+            where: {
+              username: username
+            }
+        });
+        const userGroupNames = userGroups.map(userGroup => userGroup.groupName);
+
+        const allGroups = await Group.findAll({ order: [['name', 'ASC']] });
+        const groupsJSON = await Promise.all(allGroups.map(async (group) => {
+          const userCount = await UserGroup.count({
+              where: {
+                  groupName: group.name
+              }
+          });
+          return {
+              name: group.name,
+              isMember: userGroupNames.includes(group.name),
+              isFull: userCount === 20
+          };
+        }));
+
+        res.json({ groups: groupsJSON });
+
+    } catch (error) {
+      return res.status(500).json({ error: 'Internal Server Error' });
+    }
+});
+
+
+
+// Getting a group by its name
+router.get('/:name', async (req, res) => {
+    try {
+        const groupName = req.params.name;
+
+        // Need also to get the group members
+        const group = await Group.findOne({
+            where: {
+                name: groupName
+            }
+        });
+        if (!group) {
+            return res.status(404).json({ error: 'Group not found' });
+        }
+
+        const userGroups = await UserGroup.findAll({
+            where: {
+              groupName: groupName
+            }
+        });
+
+        // Construct JSON response
+        const groupJSON = group.toJSON();
+        groupJSON.users = userGroups.map(userGroup => userGroup.username);
+
+        res.json(groupJSON);
+    } catch (error) {
+        return res.status(400).json({ error: error.message });
+    }
+});
+
 
 //Api middleware
 router.use('/api', apiRoutes);
